+++ NEWS.md
@@ -6,6 +6,10 @@ Version 0.17.1
  - Check attribute service validity at MapConfig creation time (#131)
  - Use read-only transactions for torque tiles and attributes (#130)
 
+ Enhancements:
+
+ - Allow specifying name of torque geometry column in the MapConfig (#125)
+
 Version 0.17.0
 2014-02-04